增加股票操作

+++ Udebug.pas
@@ -3,17 +3,29 @@ unit Udebug;
 interface
 
 uses
-  Windows, Classes, SysUtils, Variants, adxAddIn, forms, FileCtrl,
-  excel2000, StrUtils, ZcGridClasses, ZJGrid, ZcDataGrid, ZcDBGrids, ZcUniClass,
-  Dialogs, Controls, IniFiles, ShellAPI, communit,
-  ExtCtrls, DB, ADODB;
+  Windows, Classes, SysUtils, Variants, adxAddIn, forms, FileCtrl, excel2000,
+  StrUtils, ZcGridClasses, ZJGrid, ZcDataGrid, ZcDBGrids, ZcUniClass, Dialogs,
+  Controls, IniFiles, ShellAPI, ExtCtrls, DB, ADODB;
+
+const
+  WilltoDebug = true;
+
+var
+  mainpath: string;
 
 procedure NewTxt(FileName: string);
+
 procedure OpenTxt(FileName: string);
+
 procedure ReadTxt(FileName: string);
+
 procedure AppendTxt(Str: string; FileName: string);
+
 procedure debugto(str: string);
+
 procedure DebugReset();
+
+procedure DebugList;
 
 implementation
 
@@ -21,6 +33,8 @@ procedure NewTxt(FileName: string);
 var
   F: Textfile;
 begin
+  if not WilltoDebug then
+    exit;
   if fileExists(FileName) then
     DeleteFile(FileName); {看文件是否存在,在就刪除}
   AssignFile(F, FileName); {将文件名与变量 F 关联}
@@ -33,6 +47,8 @@ procedure OpenTxt(FileName: string);
 var
   F: Textfile;
 begin
+  if not WilltoDebug then
+    exit;
   AssignFile(F, FileName); {将文件名与变量 F 关联}
   Append(F); {以编辑方式打开文件 F }
   Writeln(F, '将您要写入的文本写入到一个 .txt 文件');
@@ -44,6 +60,8 @@ var
   F: Textfile;
   str: string;
 begin
+  if not WilltoDebug then
+    exit;
   AssignFile(F, FileName); {将文件名与变量 F 关联}
   Reset(F); {打开并读取文件 F }
   Readln(F, str);
@@ -55,6 +73,8 @@ procedure AppendTxt(Str: string; FileName: string);
 var
   F: Textfile;
 begin
+  if not WilltoDebug then
+    exit;
   AssignFile(F, FileName);
   Append(F);
   Writeln(F, Str);
@@ -63,6 +83,8 @@ end;
 
 procedure DebugReset();
 begin
+  if not WilltoDebug then
+    exit;
   NewTxt(mainpath + 'test.txt');
 end;
 
@@ -70,6 +92,8 @@ procedure debugto(str: string);
 var
   afile: string;
 begin
+  if not WilltoDebug then
+    exit;
   afile := mainpath + 'test.txt';
 
   if not FileExists(afile) then
@@ -77,6 +101,13 @@ begin
 
   AppendTxt(Str, afile);
 
+end;
+
+procedure DebugList;
+begin
+  if not WilltoDebug then
+    exit;
+  ShellExecute(0, 'open', PChar(mainpath + 'TEST.txt'), '', nil, 1);
 end;
 
 end.
